requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
email-validator
Werkzeug
python-dotenv
SQLAlchemy
PyMySQL
